refactor: separate page.tsx into page,screen.tsx like MVVM
