+++ apps/web/app/library/page.tsx
@@ -1,9 +1,5 @@
-import PageLayout from "@/_components/layout/PageLayout";
+import LibraryScreen from "./screen";
 
 export default function Library() {
-  return (
-    <PageLayout>
-      <h1 className="text-4xl font-bold text-white w-full">Library</h1>
-    </PageLayout>
-  );
+  return <LibraryScreen />;
 }
