+++ apps/web/app/library/screen.tsx
@@ -1,0 +1,9 @@
+import PageLayout from "@/_components/layout/PageLayout";
+
+export default function LibraryScreen() {
+  return (
+    <PageLayout>
+      <h1 className="text-4xl font-bold text-white w-full">Library</h1>
+    </PageLayout>
+  );
+}
