+++ apps/web/app/mypage/screen.tsx
@@ -1,0 +1,9 @@
+import PageLayout from "@/_components/layout/PageLayout";
+
+export default function MyPageScreen() {
+  return (
+    <PageLayout>
+      <h1 className="text-4xl font-bold text-white w-full">My Page</h1>
+    </PageLayout>
+  );
+}
