+++ apps/web/app/page.tsx
@@ -1,10 +1,5 @@
-import React from "react";
-import PageLayout from "@/_components/layout/PageLayout";
+import HomeScreen from "./screen";
 
 export default function Home() {
-  return (
-    <PageLayout>
-      <h1 className="text-4xl font-bold text-white w-full">Home</h1>
-    </PageLayout>
-  );
+  return <HomeScreen />;
 }
