+++ apps/web/app/screen.tsx
@@ -1,0 +1,9 @@
+import PageLayout from "@/_components/layout/PageLayout";
+
+export default function HomeScreen() {
+  return (
+    <PageLayout>
+      <h1 className="text-4xl font-bold text-white w-full">Home</h1>
+    </PageLayout>
+  );
+}
